replaced login and register
