+++ src/components/NonnaLogin.js
@@ -1,117 +1,129 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-import { Stack } from '@mui/material';
-import { Button, Container, TextField } from '@material-ui/core';
-import InputAdornment from '@mui/material/InputAdornment';
-import IconButton from '@mui/material/IconButton';
-import Visibility from '@mui/icons-material/Visibility';
-import VisibilityOff from '@mui/icons-material/VisibilityOff';
+import * as React from 'react';
+import Avatar from '@mui/material/Avatar';
+import Button from '@mui/material/Button';
+import TextField from '@mui/material/TextField';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import Checkbox from '@mui/material/Checkbox';
+import Link from '@mui/material/Link';
+import Paper from '@mui/material/Paper';
+import Box from '@mui/material/Box';
+import Grid from '@mui/material/Grid';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+import Typography from '@mui/material/Typography';
 import { orange } from '@mui/material/colors';
 import { styled } from '@mui/material/styles';
-import { Link } from 'react-router-dom';
 
 export default function NonnaLogin() {
-    const classes = useStyles();
-    const handleMouseDownPassword = (event) => {
-        event.preventDefault();
-    };
-
-    const handleChange = (prop) => (event) => {
-        setValues({ ...values, [prop]: event.target.value });
-    };
-
-    const handleClickShowPassword = () => {
-        setValues({
-            ...values,
-            showPassword: !values.showPassword,
-        });
-    };
-
-    const [values, setValues] = React.useState({
-        password: '',
-        showPassword: false,
+  const handleSubmit = (event) => {
+    event.preventDefault();
+    const data = new FormData(event.currentTarget);
+    console.log({
+      email: data.get('email'),
+      password: data.get('password'),
     });
+  };
 
-    return (
-        <div className={classes.root}>
-            <Stack>
-                <Container>
-                    <Typography className={classes.loginTitle} variant="h4">
-                        Inicia sesión
-                    </Typography>
-                </Container>
-                <Container>
-                    <TextField
-                        fullWidth
-                        variant="outlined"
-                        label="Correo electrónico"
-                        className={classes.searchInput}
-                    />
-                    <TextField
-                        id="outlined-adornment-password"
-                        className={classes.searchInput}
-                        fullWidth
-                        variant="outlined"
-                        type={values.showPassword ? 'text' : 'password'}
-                        value={values.password}
-                        label="Contraseña"
-                        onChange={handleChange('password')}
-                        InputProps={{
-                            endAdornment: (
-                                <InputAdornment position="end">
-                                    <IconButton
-                                        aria-label="Mostrar/Ocultar"
-                                        onClick={handleClickShowPassword}
-                                        onMouseDown={handleMouseDownPassword}
-                                        edge="end"
-                                    >
-                                        {values.showPassword ? <Visibility /> : <VisibilityOff />}
-                                    </IconButton>
-                                </InputAdornment>
-                            ),
-                        }}
-                    />
-                </Container>
-                <Container>
-                    <Stack className={classes.loginButtons}>
-                        <LoginButton>Ingresar</LoginButton>
-                        <Link to="/olvido-de-password" color="inherit">Olvidaste la contraseña?</Link>
-                        <Link to="/registro" color="inherit">No tenés cuenta? Registrate!</Link>
-                    </Stack>
-                </Container>
-            </Stack>
-        </div>
-    );
+  return (
+    <Grid container component="main" sx={{ height: '100vh' }}>
+      <Grid
+        item
+        xs={false}
+        sm={4}
+        md={7}
+        sx={{
+          backgroundImage: 'url(https://images.unsplash.com/photo-1555939594-58d7cb561ad1?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80)',
+          backgroundRepeat: 'no-repeat',
+          backgroundColor: (t) =>
+            t.palette.mode === 'light' ? t.palette.grey[50] : t.palette.grey[900],
+          backgroundSize: 'cover',
+          backgroundPosition: 'center',
+        }}
+      />
+      <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
+        <Box
+          sx={{
+            my: 8,
+            mx: 4,
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+          }}
+        >
+          <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+            <LockOutlinedIcon />
+          </Avatar>
+          <Typography component="h1" variant="h5">
+            Iniciar Sesión
+          </Typography>
+          <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 1 }}>
+            <TextField
+              margin="normal"
+              required
+              fullWidth
+              id="email"
+              label="Correo Electronico"
+              name="email"
+              autoComplete="email"
+              autoFocus
+            />
+            <TextField
+              margin="normal"
+              required
+              fullWidth
+              name="password"
+              label="Contrasena"
+              type="password"
+              id="password"
+              autoComplete="current-password"
+            />
+            <FormControlLabel
+              control={<Checkbox value="remember" color="primary" />}
+              label="Recordame"
+            />
+            <LoginButton
+              type="submit"
+              fullWidth
+              variant="contained"
+              sx={{ mt: 3, mb: 2 }}>
+              Iniciar sesión
+            </LoginButton>
+            <Grid container>
+              <Grid item xs>
+                <Link href="/olvido-de-password" variant="body2">
+                  Olvidaste la contrasena?
+                </Link>
+              </Grid>
+              <Grid item>
+                <Link href="/registro" variant="body2">
+                  No tenés cuenta? Registrate!
+                </Link>
+              </Grid>
+            </Grid>
+            <Copyright sx={{ mt: 5 }} />
+          </Box>
+        </Box>
+      </Grid>
+    </Grid>
+  );
 }
 
-const useStyles = makeStyles(() => ({
-    root: {
-        marginLeft: "100px",
-        marginRight: "100px"
-    },
-    loginTitle: {
-        textAlign: "center",
-        marginBottom: "50px"
-    },
-    searchInput: {
-        color: "black",
-        borderBottomColor: 'white',
-        marginBottom: "25px",
-    },
-    loginButtons: {
-        alignItems: "center"
-    }
+const LoginButton = styled(Button)(({ theme }) => ({
+  color: theme.palette.getContrastText(orange[500]),
+  backgroundColor: orange[500],
+  '&:hover': {
+    backgroundColor: orange[700],
+  },
+  borderRadius: 20
 }));
 
-const LoginButton = styled(Button)(({ theme }) => ({
-    color: theme.palette.getContrastText(orange[500]),
-    backgroundColor: orange[500],
-    '&:hover': {
-        backgroundColor: orange[700],
-    },
-    borderRadius: 20,
-    paddingLeft: "15px",
-    paddingRight: "15px",
-    marginBottom: "15px"
-}));
+function Copyright(props) {
+  return (
+    <Typography variant="body2" color="text.secondary" align="center" {...props}>
+      {'Copyright © '}
+      <Link to='/' color="inherit" > La Nonna      </Link>
+      {' '}
+      {new Date().getFullYear()}
+      {'.'}
+    </Typography>
+  );
+}
