--- src/components/NonnaRegister2.js
@@ -1,135 +1,0 @@
-import * as React from 'react';
-import Avatar from '@mui/material/Avatar';
-import Button from '@mui/material/Button';
-import TextField from '@mui/material/TextField';
-import Paper from '@mui/material/Paper';
-import Box from '@mui/material/Box';
-import Grid from '@mui/material/Grid';
-import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
-import Typography from '@mui/material/Typography';
-import { orange } from '@mui/material/colors';
-import { styled } from '@mui/material/styles';
-
-export default function NonnaRegister2() {
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        const data = new FormData(event.currentTarget);
-        console.log({
-            name: data.get('name'),
-            lastName: data.get('lastName'),
-            email: data.get('email'),
-            cel: data.get('cel'),
-            password: data.get('password'),
-            repitPassword: data.get('repitPassword')
-        });
-    };
-
-    return (
-        <Grid container component="main" sx={{ height: '100vh' }}>
-            <Grid
-                item
-                xs={false}
-                sm={4}
-                md={7}
-                sx={{
-                    backgroundImage: 'url(https://images.unsplash.com/photo-1555939594-58d7cb561ad1?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80)',
-                    backgroundRepeat: 'no-repeat',
-                    backgroundColor: (t) =>
-                        t.palette.mode === 'light' ? t.palette.grey[50] : t.palette.grey[900],
-                    backgroundSize: 'cover',
-                    backgroundPosition: 'center',
-                }}
-            />
-            <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
-                <Box
-                    sx={{
-                        my: 8,
-                        mx: 4,
-                        display: 'flex',
-                        flexDirection: 'column',
-                        alignItems: 'center',
-                    }}
-                >
-                    <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
-                        <LockOutlinedIcon />
-                    </Avatar>
-                    <Typography component="h1" variant="h5">
-                        Registrate!
-                    </Typography>
-                    <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 1 }}>
-                        <TextField
-                            margin="normal"
-                            required
-                            fullWidth
-                            id="name"
-                            name="name"
-                            label="Nombre"
-                            autoFocus
-                        />
-                        <TextField
-                            margin="normal"
-                            required
-                            fullWidth
-                            id="lastName"
-                            name="lastName"
-                            label="Apellido"
-                        />
-                        <TextField
-                            margin="normal"
-                            required
-                            fullWidth
-                            id="email"
-                            name="email"
-                            label="Correo Electronico"
-                            autoComplete="email"
-                        />
-                        <TextField
-                            margin="normal"
-                            required
-                            fullWidth
-                            id="cel"
-                            name="cel"
-                            type="number"
-                            label="Teléfono"
-                        />
-                        <TextField
-                            margin="normal"
-                            required
-                            fullWidth
-                            id="password"
-                            name="password"
-                            label="Contraseña"
-                            type="password"
-                            autoComplete="current-password"
-                        />
-                        <TextField
-                            margin="normal"
-                            required
-                            fullWidth
-                            id="repitPassword"
-                            name="repitPassword"
-                            label="Repetir Contraseña"
-                            type="password"
-                        />
-                        <RegisterButton
-                            type="submit"
-                            fullWidth
-                            variant="contained"
-                            sx={{ mt: 3, mb: 2 }}>
-                            Registrate
-                        </RegisterButton>
-                    </Box>
-                </Box>
-            </Grid>
-        </Grid>
-    );
-}
-
-const RegisterButton = styled(Button)(({ theme }) => ({
-    color: theme.palette.getContrastText(orange[500]),
-    backgroundColor: orange[500],
-    '&:hover': {
-        backgroundColor: orange[700],
-    },
-    borderRadius: 20
-}));
